requirements.txt
streamlit==1.40.0
langchain
openai
langchain-community
pydantic
langchain_openai
atlassian-python-api